@@ -36,8 +36,8 @@
     "prop-types": "15.6.0",
     "querystringify": "1.0.0",
     "ramda": "^0.25.0",
-    "react": "16.3.1",
-    "react-native": "0.55.1",
+    "react": "^16.4.2",
+    "react-native": "^0.56.0",
     "react-native-animatable": "^1.3.0",
     "react-native-audioplayer": "^0.2.0",
     "react-native-camera": "^1.2.0",
@@ -50,8 +50,8 @@
     "react-native-sound": "^0.10.9",
     "react-native-tensorflow": "^0.1.8",
     "react-native-vector-icons": "^5.0.0",
-    "react-navigation": "1.0.0-beta.11",
-    "react-navigation-redux-helpers": "^1.0.4",
+    "react-navigation": "^2.11.2",
+    "react-navigation-redux-helpers": "^2.0.5",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2",
     "redux-persist": "^5.4.0",
@@ -65,12 +65,12 @@
     "@storybook/addons": "^3.4.0",
     "@storybook/channels": "^3.4.0",
     "@storybook/react-native": "^3.4.0",
-    "babel-core": "6.26.0",
+    "babel-core": "^6.26.3",
     "babel-eslint": "7.1.1",
     "babel-jest": "23.4.2",
     "babel-plugin-ignite-ignore-reactotron": "^0.3.0",
     "babel-preset-env": "^1.6.1",
-    "babel-preset-react-native": "^3.0.2",
+    "babel-preset-react-native": "5.0.0",
     "enzyme": "^3.2.0",
     "enzyme-adapter-react-16": "^1.1.0",
     "husky": "^0.14.3",
