@@ -7,7 +7,6 @@ import createStore from '../Redux'
 
 // create our store
 const store = createStore()
-
 /**
  * Provides an entry point into our application.  Both index.ios.js and index.android.js
  * call this component first.
@@ -18,6 +17,11 @@ const store = createStore()
  * We separate like this to play nice with React Native's hot reloading.
  */
 class App extends Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+    }
+  }
   render () {
     return (
       <Provider store={store}>
